requirements.txt
absl-py>=1.0.0
astunparse==1.6.3
autocorrect==2.6.1
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.4
colorama==0.4.6
cycler==0.12.1
decorator==4.4.2
Flask==1.1.1
flatbuffers>=2.0,<3.0
gast==0.3.3
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==2.10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.4.2
Keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.4
moviepy==1.0.3
nltk==3.9.1
numpy>=1.22,<1.27
oauthlib==3.2.2
opencv-python==4.5.3.56
opt-einsum==3.3.0
pandas==1.2.1 
pillow==10.4.0
proglog==0.1.10
protobuf==3.19.6
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==0.24.1
scipy==1.10.1
six==1.15.0
SpeechRecognition==3.10.4
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
termcolor==1.1.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.67.1
typing-extensions==3.7.4.3
urllib3==2.2.3
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.20.2
